Add a wrapper library for libcap
The `caps` crate does not support setting Effective, Inheritable, or
Permitted capabilities [1].  It also requires dynamic allocation to use
the `capget` and `capset` wrappers, which is not desired in the Zygote.

For these reasons I have decided to wrap the needed functionality from
`libcap`, as it is known to work with the existing Zygote.

1. https://docs.rs/caps/latest/src/caps/lib.rs.html#304-311

Test: m libcapwrap
Bug: 415208344
Change-Id: Iacdfff59936bfd21af3233fdcaa37c25b4bbeace