Modify the fallback socket name to pick up sockets from init
init can create sockets before fork and pass it to the child process
with the FD number encoded in the environment. We had infrastructure
to get a FD number from environment in case --socket was not specified,
but the prefix was not quite right. Change it to the one init uses.

Since the default env var name depends on the --name option, move the
logic out of the parser. Instead, provide a specimen-specific method
that determines the socket value from parsed CLI options.

Bug: 430174867
Test: native zygote starts with SELinux enforcing
Flag: build.RELEASE_NATIVE_FRAMEWORK_PROTOTYPE
Change-Id: Id6da70198e0633ecd92c9515fa63412312795db2