Add SocketAddr type to wrap both sockaddr and its length
Currently, the abstract socket functions pass `sizeof sockaddr_un` as
socklen. When hashing and comparing abstract socket names, the kernel
does not assume it is a NUL-terminated string and cares about socklen.
To interop with implementations other than zygote-sys, we need to make
socklen match the string length.

Introduce a new SocketAddr type wrapping the sockaddr and its length,
and use it from the helpers. Existing callsites are not changed because
they infer the type and they will pass around SocketAddr instead of
libc::sockaddr_un from now on.

The file descriptor watcher is also reworked so that it preserves NUL
bytes in socket name. The previous implementation would fail since
setting the correct length means that [..socklen] no longer contains a
NUL.

Bug: 443729061
Test: cargo clippy && cargo test
Change-Id: I82853f01f9540a7e1dec8bc41b1066c3dbbb2732