Make zygote_next startup timing configurable
In Android, zygote_next is currently started in the boot process like
the managed zygote but this CL makes it configurable based on a
property.
If `persist.zygote.zygote_next.start_on_boot` is true, then the service
is started by init automatically but otherwise the service won't be
started unless someone explicitly starts it.

And also this CL makes zygote_next set a property to notify that the
server is ready not to cause a race between service status and status of
the socket listening to requests from clients.

Bug: 452150691
Test: atest NativeZygoteIntegrationTest
Test: atest zygote_next_unit_tests zygote_next_integration_test_fd_no_panic
Flag: build.RELEASE_NATIVE_FRAMEWORK_PROTOTYPE
Change-Id: Ie485e2079f80796b8e5a85db6500acefb107aa9b