Register file descriptors used for logging in Server::new
Currently some FDs used for logging, specifically the /dev/pmsg0 file
and the socket connected to /dev/socket/logdw, are not reflected to the
FileDescriptorRegistry.

This commit makes sure these FDs are registered by implementing a new
trait method `peer_socket_is_allowed` to get the allowlisted peer socket
name and by adding the two file paths to specific allowlists in the
android_native module. By this change `register_new` will now register
the log FDs, and we call it in `Server:new`.

Test: m
Test: Clank native renderers work with http://gpaste/4631571747438592
Bug: 457960847
Bug: 451831043
Change-Id: I2daa0d626163c36427a38bbb118089e01f49e028