Define SocketAddress to express the socket address type
The current code uses `FileDescriptorInfo::AbstractSocket` to express
abstract sockets, but we should use a different enum for the socket
address as `AbstractSocket` here is a bound socket with abstract socket
name. This commit unifies `AbstractSocket` to `BoundSocket` and let
`BoundSocket` have the newly defined `SocketAddress`, whose variants
are, for now, `Path` and `Abstract`.

This change makes the introduction of
`FileDescriptorInfo::ConnectedSocket` smoother, which will be
implemented in the next commit.

Test: Clank native renderers work with http://gpaste/4631571747438592
Bug: 457960847
Bug: 451831043
Change-Id: I2bc1eca6fdcca8c4811697782c31b620e74f4414