Set zygote_next priority to -20
The Native Zygote spends most of its time asleep, only waking up in
response to process creating requests from the System Server.  During
this CUJ the Native Zygote should be woken up as soon as possible and
executed with the highest priority.

Running the Native Zygote with a -20 priority will match the behavior of
the Managed Zygote.

Test: atest CtsNativeServiceTest
Change-Id: I9183ed007eb7d8aa1e78b5d1a698265fc97a8db7